More cleanup, cleans up tests, cleans up the BuildManagerAssemblyProvider (even though it's not used)
