+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -13,7 +13,6 @@
     <ItemGroup>
       <PackageReference Include="System.ComponentModel.Annotations" Version="4.6.0" />
       <PackageReference Include="System.Reflection.Emit.Lightweight" Version="4.3.0" />
-      <PackageReference Include="System.Reflection.Metadata" Version="1.8.0" />
       <PackageReference Include="System.Runtime.Caching" Version="4.6.0" />
     </ItemGroup>
