+++ src/Umbraco.Tests/Cache/HttpRequestAppCacheTests.cs
@@ -16,7 +16,7 @@ namespace Umbraco.Tests.Cache
         public override void Setup()
         {
             base.Setup();
-            var typeFinder = new TypeFinder(Mock.Of<ILogger>(), new DefaultUmbracoAssemblyProvider(GetType().Assembly));
+            var typeFinder = TestHelper.GetTypeFinder();
             _ctx = new FakeHttpContextFactory("http://localhost/test");
             _appCache = new HttpRequestAppCache(() => _ctx.HttpContext.Items, typeFinder);
         }
