+++ src/Umbraco.Tests/TestHelpers/BaseUsingSqlCeSyntax.cs
@@ -38,7 +38,7 @@ namespace Umbraco.Tests.TestHelpers
 
             var ioHelper = TestHelper.IOHelper;
             var logger = new ProfilingLogger(Mock.Of<ILogger>(), Mock.Of<IProfiler>());
-            var typeFinder = new TypeFinder(Mock.Of<ILogger>(), new DefaultUmbracoAssemblyProvider(GetType().Assembly));
+            var typeFinder = TestHelper.GetTypeFinder();
             var typeLoader = new TypeLoader(ioHelper, typeFinder, NoAppCache.Instance,
                 new DirectoryInfo(ioHelper.MapPath("~/App_Data/TEMP")),
                 logger,
