+++ src/Umbraco.Tests/Web/UmbracoHelperTests.cs
@@ -28,7 +28,7 @@ namespace Umbraco.Tests.Web
         {
             // FIXME: bad in a unit test - but Udi has a static ctor that wants it?!
             var container = new Mock<IFactory>();
-            var typeFinder = new TypeFinder(Mock.Of<ILogger>(), new DefaultUmbracoAssemblyProvider(GetType().Assembly));
+            var typeFinder = TestHelper.GetTypeFinder();
             var ioHelper = TestHelper.IOHelper;
             container
                 .Setup(x => x.GetInstance(typeof(TypeLoader)))
