+++ src/Umbraco.Web/Composing/BuildManagerTypeFinder.cs
@@ -1,19 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Security;
 using System.Web.Compilation;
 using Umbraco.Core.Configuration;
 using Umbraco.Core.Composing;
 using Umbraco.Core.Configuration.UmbracoSettings;
-using Umbraco.Core.Hosting;
-using Umbraco.Core.IO;
 using Umbraco.Core.Logging;
 
 namespace Umbraco.Web.Composing
 {
+
     /// <summary>
     /// An implementation of TypeFinder that uses the BuildManager to resolve references for aspnet framework hosted websites
     /// </summary>
@@ -24,59 +20,12 @@ namespace Umbraco.Web.Composing
     {
 
         public BuildManagerTypeFinder(
-            IIOHelper ioHelper,
-            IHostingEnvironment hostingEnvironment,
             ILogger logger,
             IAssemblyProvider assemblyProvider,
             ITypeFinderConfig typeFinderConfig = null) : base(logger, assemblyProvider, typeFinderConfig)
         {
-            if (ioHelper == null) throw new ArgumentNullException(nameof(ioHelper));
-            if (hostingEnvironment == null) throw new ArgumentNullException(nameof(hostingEnvironment));
             if (logger == null) throw new ArgumentNullException(nameof(logger));
-
-            _allAssemblies = new Lazy<HashSet<Assembly>>(() =>
-            {
-                var isHosted = hostingEnvironment.IsHosted;
-                try
-                {
-                    if (isHosted)
-                    {
-                        var assemblies = new HashSet<Assembly>(BuildManager.GetReferencedAssemblies().Cast<Assembly>());
-
-                        //here we are trying to get the App_Code assembly
-                        var fileExtensions = new[] { ".cs", ".vb" }; //only vb and cs files are supported
-                        var appCodeFolder = new DirectoryInfo(ioHelper.MapPath(ioHelper.ResolveUrl("~/App_code")));
-                        //check if the folder exists and if there are any files in it with the supported file extensions
-                        if (appCodeFolder.Exists && fileExtensions.Any(x => appCodeFolder.GetFiles("*" + x).Any()))
-                        {
-                            try
-                            {
-                                var appCodeAssembly = Assembly.Load("App_Code");
-                                if (assemblies.Contains(appCodeAssembly) == false) // BuildManager will find App_Code already
-                                    assemblies.Add(appCodeAssembly);
-                            }
-                            catch (FileNotFoundException ex)
-                            {
-                                //this will occur if it cannot load the assembly
-                                logger.Error(typeof(TypeFinder), ex, "Could not load assembly App_Code");
-                            }
-                        }
-
-                        return assemblies;
-                    }
-                }
-                catch (InvalidOperationException e)
-                {
-                    if (e.InnerException is SecurityException == false)
-                        throw;
-                }
-
-                // Not hosted, just use the default implementation
-                return new HashSet<Assembly>(base.AssembliesToScan);
-            });
         }
-
-        private readonly Lazy<HashSet<Assembly>> _allAssemblies;
 
         /// <summary>
         /// Explicitly implement and return result from BuildManager
@@ -85,10 +34,6 @@ namespace Umbraco.Web.Composing
         /// <returns></returns>
         Type ITypeFinder.GetTypeByName (string name) => BuildManager.GetType(name, false);
 
-        /// <summary>
-        /// Explicitly implement and return result from BuildManager
-        /// </summary>
-        IEnumerable<Assembly> ITypeFinder.AssembliesToScan => _allAssemblies.Value;
 
         /// <summary>
         /// TypeFinder config via appSettings
